add app/iterwriter/options.go package
